[09.24] change source directory, add navigator
