+++ screens/SignUp.js
@@ -1,0 +1,19 @@
+import React, { Component } from 'react';
+import { StyleSheet, View, Text} from 'react-native';
+
+export default class SignUp extends Component {
+
+    render() {
+
+        return (
+            <View style={styles.container}>
+                <Text>hey!</Text>
+            </View>
+        )
+    }
+}
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+    },
+});
